Pagination for the happypath page
