+++ packages/web/src/app/scan/ScanReport.tsx
@@ -1,17 +1,25 @@
 import styled from "styled-components";
 
-import { useAppDispatch, useAppSelector } from "./store";
 import LogMessages from "../../features/logging/LogMessages";
+import { TabContainer } from "../../new-components/Tabs";
+
+import { useAppDispatch, useAppSelector } from "./store";
 import { State, changeTab } from "./slice";
+import { loadHappyPathPage } from "./slice";
+
+import { HappyPathCard } from "./HappyPathCard";
 import { ScanSummary } from "./ScanSummary";
 import ScanIssues from "./ScanIssues";
-import { TabContainer } from "../../new-components/Tabs";
-import { HappyPathCard } from "./HappyPathCard";
+import Paginator from "./Paginator";
 
 export default function ScanReport() {
   const dispatch = useAppDispatch();
 
   const { scanReport, tab, happyPathPage } = useAppSelector((state) => state.scan);
+
+  const handlePageChange = (pageIndex: number) => {
+    dispatch(loadHappyPathPage(pageIndex));
+  };
 
   if (scanReport === undefined) {
     return (
@@ -37,6 +45,12 @@ export default function ScanReport() {
               {happyPathPage.items.map((entry, index) => (
                 <HappyPathCard defaultCollapsed={true} report={entry} key={index} />
               ))}
+
+              <Paginator
+                current={happyPathPage.current}
+                total={happyPathPage.pages}
+                onPageChange={handlePageChange}
+              />
             </Summary>
           ),
         },
